split two matching methods into different files
